Initial basic action cost management
Enable to associate cost (list of transactions) to an action.
Transaction is composed of a quantity, a resource and a level (player or token).
When an action is played, the resources of player or selected token (depending on transaction level) are decreased according to each transaction.
When an action is not payable, it's deactivate and player can not do it.